requirements.txt
alembic==1.8.1
bcrypt==4.0.1
blinker==1.5
cffi==1.15.1
click==8.1.3
cryptography==38.0.4
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
idna==3.4
importlib-metadata==5.1.0
importlib-resources==5.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
paramiko==2.12.0
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==0.21.0
six==1.16.0
SQLAlchemy==1.4.44
typing_extensions==4.4.0
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.11.0
